rpm-testing: fix box and add install.sh
The original goal of this commit was to upgrade the `rpm-testing` box
provisioning shell script to be compatible with AM17 packages. This box is
useful for local testing of rpm packages, either by devs or in CI workflows.

During this endeavor we've added some other fixes to make this process easier:

- Add integration with vagrant-vbguest so the `centos/7` box supports vboxsf
  shared folders and private networking. I've also updated `Vagrantfile`
  accordingly.

- Add new `install.sh` script to provision the box but also flexible enough so
  it can be used in other contexts. This new script supports the indexless
  mode, both local and rpm remote repos, firewalld, environment string lookups
  and SELinux configuration.