Ubuntu: Configure ucf for files in /etc/default/
The apt tool sometimes asks whether to keep the current version of the
/etc/default conf files when upgrading archivematica packages, even
without having modified these files.

In this PR these files are added to ucf (Update Configuration File) so
on an upgrade or a reconfiguration the system will only ask for the
action to be taken when those files have actually been changed.