AtoM Ubuntu 20.04 LTS base box
Add Ubuntu 20.04 base box in am-packbuild for AtoM.

Note that this is using the legacy server ISO. The legacy ISO still uses
the Debian installer so is compatible with the 18.04 preseed.cfg files.
Going forward this will be removed and replaced with Subiquity installer
which will require re-implmenting the scripts.