rpm: Remove font-awesome's symlink on dashboard
When creating static content there's a broken symlink for font-awesome, that
points to `../../../frontend/node_modules/font-awesome/`.

https://github.com/artefactual/archivematica/blob/qa/1.x/src/dashboard/src/media/vendor/font-awesome

The rpm dashboard package has a different directory structure than
archivematica source code, and the `frontend` directory is not included in the
package.

This commit removes the symlink and copies the `awesome-font` directory from
the `frontend` directory.