Use unified Archivematica requirements in RPM
This commit updates the Archivematica RPM spec file so it installs the
unified virtual environment described in
https://github.com/artefactual/archivematica/issues/573.

A new `archivematica` package is introduced, which includes the
virtualenv. The other packages depend on it. In future releases, we
could stop releasing archivematica-{dashboard,mcpserver,mcpclient}
subpackages.

Connects to https://github.com/artefactual/archivematica/issues/573.

Co-authored-by: Jesús García Crespo <jesus@sevein.com>
Co-authored-by: Miguel Angel Medinilla Luque <mamedin@gmail.com>