am/deb: use virtualenv in dashboard postinst
postinst in archivematica-dashboard (Xenial and Bionic debs) were
installing all requirements globally to run Django's management tasks.
This commit updates the script so it uses the bundled virtual
environment instead. This is not a new approach, it's what we're doing
in other packages.

Connects to https://github.com/artefactual-labs/am-packbuild/issues/128.