Update AM RPM packages to use Python src layout
This updates the Archivematica and Storage Service RPM packages to
install the codebase in editable mode, removing the need to set custom
PYTHONPATHs or copy code into /usr/ directories which aligns the
packaging approach with the Ansible role for improved consistency and
maintainability.

It also adds a test and a GitHub workflow job to verify RPM package
installation via the Ansible role.